--- utils/losses/daily_reward_lin_regr_loss.py
@@ -1,104 +1,0 @@
-import torch
-from torch import nn
-from utils.losses.loss_helpers import probability_correct_class, log_class_prob_at_t_plus_zt
-
-MU_DEFAULT = 150.
-NB_DAYS_IN_YEAR = 365.
-
-class DailyRewardLinRegrLoss(nn.Module):
-    def __init__(self, alpha:float=1., weight=None, alpha_decay: list=None, epochs: int=100, start_decision_head_training: int=0, **kwargs):
-        """_summary_
-
-        Args:
-            alpha (float, optional): _description_. Defaults to 0.5.
-            weight (list, optional): weight for each class, shape: (nclasses). Defaults to None.
-            alpha_decay (list, optional): contains [alpha_decay_max, alpha_decay_min]. Through the epochs, starts at alpha_decay_max
-                        and get closer to alpha_decay_min. Defaults to None.
-            epochs (int, optional): number of epochs. Defaults to 100.
-            start_decision_head_training (int, optional): epoch to start training the decision head. Defaults to 0.
-        """
-        super(DailyRewardLinRegrLoss, self).__init__()
-
-        self.weight = weight
-        self.negative_log_likelihood = nn.NLLLoss(reduction="none", weight=self.weight)
-        self.alpha = alpha
-        
-        if alpha_decay is not None:
-            self.alpha_decay_max = alpha_decay[0]
-            self.alpha_decay_min = alpha_decay[1]
-            self.epochs = epochs
-        
-        self.start_decision_head_training = start_decision_head_training
-        # mus is a tensor of length nclasses, containing the mu for each class
-        self.mu = kwargs.get("sequencelength", NB_DAYS_IN_YEAR)*MU_DEFAULT/NB_DAYS_IN_YEAR
-        self.mus = kwargs.get("mus", torch.ones(len(weight))*self.mu).to(weight.device)
-        self.percentage_earliness_reward = torch.tensor(kwargs.get("percentage_earliness_reward", 0.5), device=weight.device)
-
-    def forward(self, log_class_probabilities, timestamps_left, y_true, return_stats=False, **kwargs):
-        N, T, C = log_class_probabilities.shape
-        epoch = kwargs.get("epoch", 0)
-
-        # equation 4, right term
-        t = torch.ones(N, T, device=log_class_probabilities.device) * \
-                torch.arange(T).type(torch.FloatTensor).to(log_class_probabilities.device)
-        log_class_probabilities_at_t_plus_zt = log_class_prob_at_t_plus_zt(log_class_probabilities, timestamps_left)
-
-        # compute alpha
-        if hasattr(self, "alpha_decay_max"):
-            if epoch >= self.start_decision_head_training:
-                # alpha goes from alpha_decay_max to alpha_decay_min linearly
-                self.alpha = self.alpha_decay_min + (self.alpha_decay_max - self.alpha_decay_min) * \
-                    (1 - (epoch-self.start_decision_head_training)/(self.epochs-self.start_decision_head_training))
-                            
-        # earliness reward and linear regression loss
-        if epoch>=self.start_decision_head_training and self.alpha<1.-1e-8:
-            earliness_reward = probability_correct_class(log_class_probabilities_at_t_plus_zt, y_true, weight=self.weight) * (1-t/T) * (1-timestamps_left.float()/T)
-            earliness_reward = earliness_reward.sum(1).mean(0)
-            
-            lin_regr_zt_loss = lin_regr_zt(t, T, self.mus, timestamps_left.float(), y_true) # (N, T)
-            lin_regr_zt_loss = lin_regr_zt_loss.sum(1).mean(0) # sum over time, mean over batch
-        else:
-            # if the decision head is not trained, the earliness reward is zero
-            earliness_reward = torch.tensor(0.0, device=log_class_probabilities.device) 
-            lin_regr_zt_loss = torch.tensor(0.0, device=log_class_probabilities.device)
-            
-        # classification loss
-        cross_entropy = self.negative_log_likelihood(log_class_probabilities.view(N*T,C), y_true.view(N*T)).view(N,T)
-        classification_loss = cross_entropy.sum(1).mean(0)
-
-        # final loss
-        alpha_1 = (1.-self.alpha)*self.percentage_earliness_reward
-        self.alphas = torch.tensor([self.alpha, alpha_1, 1.-self.alpha-alpha_1], device=log_class_probabilities.device)
-        loss = self.alphas[0]*classification_loss - self.alphas[1]*earliness_reward + self.alphas[2]*lin_regr_zt_loss
-
-        if return_stats:
-            stats = dict(
-                classification_loss=classification_loss.cpu().detach().numpy(),
-                earliness_reward=earliness_reward.cpu().detach().numpy(),
-                timestamps_left=timestamps_left.cpu().detach().numpy(),
-                lin_regr_zt_loss=lin_regr_zt_loss.cpu().detach().numpy(),
-            )
-            return loss, stats
-        else:
-            return loss
-        
-    def update_mus(self, mus):
-        self.mus = mus.clone().detach().to(device=self.mus.device)
-        
-
-def lin_regr_zt(t, T, mus, z_t, y_true):
-    """ computes the piecewise linear regression loss for z_t
-    
-    INPUT:
-    t: tensor of shape (N, T)
-    T: float 
-    mus: tensor of shape (C)
-    z_t: tensor of shape (N, T)
-    y_true: tensor of shape (N, T)
-    
-    OUTPUT: 
-    loss: tensor of shape (N, T), 
-    """
-    # lin_term is either mus[y_true]-t-z_t for t<=mus[y_true], or z_t otherwise
-    lin_term = torch.where(t<=mus[y_true], mus[y_true]-t-z_t, z_t)
-    return (lin_term/T)**2
